pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lumberjack_sdk"
version = "2.0.0dev8"
authors = [
    { name = "George Mayer", email = "george@treebeardhq.com" },
]
description = "The Lumberjack Python SDK"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "termcolor>=2.0.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-instrumentation-logging>=0.48b0",
    "opentelemetry-semantic-conventions>=0.48b0"
]

[project.optional-dependencies]
flask = [
    "opentelemetry-instrumentation-flask>=0.48b0",
]
django = [
    "opentelemetry-instrumentation-django>=0.48b0",
]
fastapi = [
    "opentelemetry-instrumentation-fastapi>=0.45b0",
]
local-server = [
    "fastapi>=0.100.0",
    "uvicorn>=0.25.0",
    "websockets>=12.0",
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.27.0",
    "opentelemetry-exporter-otlp-proto-http>=1.36.0",
    "opentelemetry-proto>=1.27.0",
    "mcp>=1.0.0",
    "requests>=2.28.0",
    "psutil>=7.0.0",
]
web = [
    "opentelemetry-instrumentation-flask>=0.48b0",
    "opentelemetry-instrumentation-django>=0.48b0",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "ruff>=0.0.290",
    "build>=1.2.2",
    "freezegun>=1.5.2",
    "opentelemetry-instrumentation-flask>=0.48b0",
    "opentelemetry-instrumentation-django>=0.48b0",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
    "types-requests>=2.32.0",
    "lumberjack-sdk[local-server]",
]

[project.scripts]
lumberjack = "lumberjack_sdk.local_server.cli:main"
lumberjack-mcp = "lumberjack_sdk.local_server.mcp_server:main"

[project.urls]
"Homepage" = "https://github.com/treebeardhq/lumberjack-sdk-python"
"Bug Tracker" = "https://github.com/treebeardhq/lumberjack-sdk-python/issues"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
markers = [
    "gevent: marks tests that require gevent",
    "eventlet: marks tests that require eventlet"
]

[tool.mypy]
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.ruff]
select = ["E", "F", "B", "I"]
line-length = 100

[tool.hatch.build.targets.wheel]
packages = ["src/lumberjack_sdk"]
include = [
    "src/lumberjack_sdk/py.typed",
    "src/lumberjack_sdk/**/*.pyi",
]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

[tool.hatch.build.targets.sdist]
include = [
    "src/lumberjack_sdk",
    "ui",
]